+++ src/bigxml/handle_mgr.py
@@ -45,7 +45,10 @@ class HandleMgr:
     def iter_from(
         self,
     ) -> Iterator["Never"]:
-        ...
+        # we don't have '...' as body in this overload as a workaround
+        # to prevent 'not-an-iterable' error in pylint
+        # see https://github.com/PyCQA/astroid/issues/1015
+        return iter(())
 
     @overload
     def iter_from(